Create  an application TestBooks
Write an application TestBooks that asks the user repeatedly for books read during summer (use a sentinel value to control the end of input). The application repeatedly creates book objects, and then prints to the screen information about the “smallest” book, (i.e. the book with the smallest number of pages), as well as the average number of pages per book
HINT: This program MUST make use of the compareTo method in order to find the “smallest” book.